Added missing keys to DecodedIdToken type and fixed other Object-like types
(1) DecodedIdToken was missing a bunch of keys which are guaranteed to
be there. The full list was copied from here:
https://firebase.google.com/docs/reference/admin/node/admin.auth.Decoded
IdToken

(2) There was an issue with some Object-like types which failed to
compile when run with strictNullChecks. See
https://groups.google.com/forum/#!msg/firebase-talk/FKtHu_3-V-I/S40fBLh-
DgAJ for more details.

I’ve verified the behavior of all of these types in VSCode and they now
behave properly.

Change-Id: Ic01f75c937a20292016c1f2d523a666966d40de2